+++ httpx/__init__.py
@@ -1,7 +1,7 @@
 from .__version__ import __description__, __title__, __version__
 from .api import delete, get, head, options, patch, post, put, request, stream
 from .auth import BasicAuth, DigestAuth
-from .client import AsyncClient, Client
+from .client import AsyncClient, Client, SyncClient
 from .config import TimeoutConfig  # For 0.8 backwards compat.
 from .config import PoolLimits, Timeout
 from .dispatch.proxy_http import HTTPProxy, HTTPProxyMode
@@ -45,6 +45,7 @@ __all__ = [
     "stream",
     "codes",
     "AsyncClient",
+    "SyncClient",
     "BasicAuth",
     "Client",
     "DigestAuth",
