+++ httpx/_async/dispatch/app.py
@@ -1,12 +1,12 @@
 import typing
 
-from ..config import TimeoutTypes
-from ..content_streams import ByteStream
-from ..models import Request, Response
-from .base import Dispatcher
+from ...config import TimeoutTypes
+from ...content_streams import ByteStream
+from ...models import Request, Response
+from .base import AsyncDispatcher
 
 
-class ASGIDispatch(Dispatcher):
+class AsyncAppDispatch(AsyncDispatcher):
     """
     A custom dispatcher that handles sending requests directly to an ASGI app.
