+++ httpx/backends/sync.py
@@ -1,0 +1,9 @@
+from .._sync.backends.base import SyncConcurrencyBackend
+
+
+class SyncBackend(SyncConcurrencyBackend):
+    """
+    A synchronous backend.
+
+    TODO
+    """
