+++ httpx/dispatch/__init__.py
@@ -2,6 +2,6 @@
 Dispatch classes handle the raw network connections and the implementation
 details of making the HTTP request and receiving the response.
 """
-from .asgi import ASGIDispatch
+from .._async.dispatch.app import AsyncAppDispatch as ASGIDispatch
 
 __all__ = ["ASGIDispatch"]
