+++ httpx/dispatch/base.py
@@ -1,52 +1,3 @@
-import typing
-from types import TracebackType
+from .._async.dispatch.base import AsyncDispatcher as Dispatcher
 
-from ..config import Timeout
-from ..models import (
-    HeaderTypes,
-    QueryParamTypes,
-    Request,
-    RequestData,
-    Response,
-    URLTypes,
-)
-
-
-class Dispatcher:
-    """
-    Base class for dispatcher classes, that handle sending the request.
-
-    Stubs out the interface, as well as providing a `.request()` convenience
-    implementation, to make it easy to use or test stand-alone dispatchers,
-    without requiring a complete `AsyncClient` instance.
-    """
-
-    async def request(
-        self,
-        method: str,
-        url: URLTypes,
-        *,
-        data: RequestData = b"",
-        params: QueryParamTypes = None,
-        headers: HeaderTypes = None,
-        timeout: Timeout = None,
-    ) -> Response:
-        request = Request(method, url, data=data, params=params, headers=headers)
-        return await self.send(request, timeout=timeout)
-
-    async def send(self, request: Request, timeout: Timeout = None) -> Response:
-        raise NotImplementedError()  # pragma: nocover
-
-    async def close(self) -> None:
-        pass  # pragma: nocover
-
-    async def __aenter__(self) -> "Dispatcher":
-        return self
-
-    async def __aexit__(
-        self,
-        exc_type: typing.Type[BaseException] = None,
-        exc_value: BaseException = None,
-        traceback: TracebackType = None,
-    ) -> None:
-        await self.close()
+__all__ = ["Dispatcher"]
