+++ scripts/clean
@@ -1,5 +1,8 @@
 #!/bin/sh -e
 
+if [ -d 'build' ] ; then
+    rm -r build
+fi
 if [ -d 'dist' ] ; then
     rm -r dist
 fi
