+++ scripts/compile
@@ -1,0 +1,13 @@
+#!/bin/sh -e
+
+export PREFIX=""
+if [ -d "venv" ]; then
+  export PREFIX="venv/bin/"
+fi
+
+set -x
+
+scripts/build
+rm -rf httpx/_sync
+mv build/lib/httpx/_sync httpx
+scripts/clean
