+++ tests/test_concurrency.py
@@ -4,8 +4,8 @@ import pytest
 import trio
 
 from httpx import Timeout
+from httpx._async.backends.base import lookup_backend
 from httpx.backends.asyncio import AsyncioBackend
-from httpx.backends.base import lookup_backend
 from httpx.backends.trio import TrioBackend
 from httpx.config import SSLConfig
 from tests.concurrency import get_cipher, run_concurrently
